+++ Changelog.md
@@ -6,6 +6,9 @@
 
 ### Ignition Rendering 4.0.0
 
+1. Scale BaseAxis properly
+    * [Pull request #88](https://github.com/ignitionrobotics/ign-rendering/pull/88)
+
 1. Add ogre2 AxisVisual and ArrowVisual
     * [Pull request 87](https://github.com/ignitionrobotics/ign-rendering/pull/87)